fix(errors): resolve remaining compilation errors
修复剩余的编译错误：

1. secure_data_service.dart: 添加dart:convert导入
2. window_service.dart: 添加flutter/foundation.dart导入
3. memory_optimizer.dart: 修复类型转换 (num → int/double)
4. global_search_widget.dart: ActionChip → InputChip
5. secure_storage_service.dart: 添加uuid导入
6. ai_service_integration.dart: 修复导出语句

+++ lib/services/performance/memory_optimizer.dart
@@ -506,11 +506,11 @@ class MemoryOptimizer {
       final estimatedUsage = _estimateMemoryUsage();
 
       return MemoryUsageStats(
-        totalMemoryMB: estimatedUsage['total'] ?? 1024,
-        usedMemoryMB: estimatedUsage['used'] ?? 100,
-        freeMemoryMB: estimatedUsage['free'] ?? 924,
-        usagePercentage: estimatedUsage['percentage'] ?? 10.0,
-        state: _determineMemoryState(estimatedUsage['percentage'] ?? 10.0),
+        totalMemoryMB: (estimatedUsage['total'] ?? 1024).toInt(),
+        usedMemoryMB: (estimatedUsage['used'] ?? 100).toInt(),
+        freeMemoryMB: (estimatedUsage['free'] ?? 924).toInt(),
+        usagePercentage: (estimatedUsage['percentage'] ?? 10.0).toDouble(),
+        state: _determineMemoryState((estimatedUsage['percentage'] ?? 10.0).toDouble()),
         timestamp: DateTime.now(),
       );
     } catch (e) {
